[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cc
@@ -2,11 +2,11 @@
 #include <string>
 using namespace std;
 
-int calculate(int A, int B) {
+long long int calculate(long int A, long int B) {
     return (A+B)*(A-B);
 }
 int main(){
-    int A,B;
+    long int A,B;
     cin >> A >> B;
     cout << calculate(A,B);
     return 0;
